Added an option to disable routing directory index
In some frameworks, passing index.php, or the directory index back from
the router would cause "/" to fail to load a page since it's passing
back the output of the directory index.

Some frameworks, like Zend Framework 1 don't behave properly when
index.php is passed back.

This adds a routeToDirectoryIndex option for config which can be set to
0 to turn off passing back the directory index. By default it's 1 to
ensure the router behaves normally for existing installations.